// src/app/models/AllPokemons.interface.ts
// Generated by https://quicktype.io

import { NonNullableFormBuilder } from "@angular/forms";

// Generated by https://quicktype.io

export interface ResponsePokemon {
    headers:         Headers;
    body:            AllPokemons[];
    statusCode:      string;
    statusCodeValue: number;
}

export interface ResponseCreation {
    headers: Headers;
    body: string;
    statusCode: string;
    statusCodeValue: number;
}


export interface Headers {
}


export interface AllPokemons {

    image:          string;
    name:           string;
    image_animated: string;
    id:             number;
    Weight:         number;
    height:         number;
    types: Type[];
}


export interface DetailPokemon {
    headers:         Headers;
    body:            Pokemon;
    statusCode:      string;
    statusCodeValue: number;
}


export interface Pokemon {
    id:        number | null;
    weight:    number;
    height:    number ;
    name:      string ;
    species:   string ;
    sprites:   Sprites ;
    abilities: Ability[];
    stats:     Stat[];
    types:     Type[];
}

export interface NewPokemon {
    weight:    number;
    height:    number ;
    name:      string ;
    species:   string ;
    sprites:   Object ;
    abilities: Object[];
    stats:     Object[];
    types:     Type[];
}

export interface Ability {
    id:   number;
    name: string;
    slot: number;
}

export interface Sprites {
    id:           number;
    backDefault:  string;
    frontDefault: string;
    versions:     string;
}

export interface Stat {
    id:       number;
    baseStat: number;
    stat:     string;
}

export interface Type {
    id:   number;
    type: string;
}


export interface responseType {
    headers:         Headers;
    body:            Type[];
    statusCode:      string;
    statusCodeValue: number;
}